Autenticacion agregada en talleres, además de la protección de las rutas relacionadas

+++ TLACUALLI/app/Http/Controllers/PublicacionesController.php
@@ -6,6 +6,8 @@ use App\Models\Usuarios;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
+use Illuminate\Support\Facades\Auth;
+
 use Carbon\Carbon;
 
 class PublicacionesController extends Controller
@@ -15,25 +17,47 @@ class PublicacionesController extends Controller
      */
     public function index()
     {
-        //Metodo temporal ya que no hay de momento roles o login
-        //Obtener todas las publicaciones de tipo taller
-        $publicaciones = Publicaciones::where('id_tipo', 2)->
-            where('estatus', true)
+        //Verifica si el usuario está autenticado, si está autenticado se le habilita la opcion
+        // para ir a sus talleres 
+        $usuarioId = Auth::id();
+
+        if (Auth::check()) {
+            $usuario = Usuarios::with('roles')
+            ->find($usuarioId);
+        }
+        else {
+            //Si no está autenticado se crea una clase generica para que pueda visualizar todos los talleres activos
+            $usuario = new \stdClass();
+            $usuario->roles = new \stdClass();
+            $usuario->roles->id = null;
+        }
+        //Busca todas las publicaciones de tipo taller, activas y con los datos del publicador
+        $publicaciones = Publicaciones::where('id_tipo', 2)
+            ->where('estatus', true)
             ->with('usuario')
             ->get();
         
         //Envia talleres a la vista de talleres
-        return view('talleres', compact('publicaciones'));
+        return view('talleres', compact('publicaciones', 'usuario'));
     }
 
     public function index_mis_talleres()
     {
-        //Metodo temporal ya que no hay de momento roles o login, utilizado para mostrar los talleres del usuario
-        //Obtener todas las publicaciones de tipo taller
-        $publicaciones = Publicaciones::where('id_tipo', 2)->with('usuario')->get();
+        $usuarioId = Auth::id();
         
-        //Envia talleres a la vista de talleres
-        return view('mis_talleres', compact('publicaciones'));
+        if (Auth::check()){
+            //Obtiene todos los talleres relacionados a este usuario
+            $publicaciones = Publicaciones::where('id_usuario', $usuarioId)
+            ->with('usuario')
+            ->get();
+            
+            //Envia talleres a la vista de talleres
+            return view('mis_talleres', compact('publicaciones'));
+        }
+        else {
+            //Si no esta autenticado lo manda al home
+            return redirect('/');
+        }   
     }
 
     /**
@@ -49,45 +73,48 @@ class PublicacionesController extends Controller
      */
     public function store(Request $request)
     {
-        //dd($request->all());
-        //Validaciones
-        $validator = $request->validate([
-            '_descT'                => 'required',
-            '_nt'                   => 'required|max:50',
-            '_contT'                => 'required|file|max:2048',
-            '_costoT'               => 'numeric',
-            //'id_usuario'            => 'required',
-            //'id_tipo'               => 'required',
-            //'id_usuario_revision'   => 'required',
-            //'fecha_revision'        => 'date',
-            //'estatus'               => 'required'
-        ]);
+        $usuarioId = Auth::id();
+        
+        if (Auth::check()){
+            //Validaciones
+            $validator = $request->validate([
+                '_descT'                => 'required',
+                '_nt'                   => 'required|max:50',
+                '_contT'                => 'required|file|max:2048', //Ruta de la imagen
+                '_costoT'               => 'numeric',
+                //'id_usuario_revision'   => 'required',
+                //'fecha_revision'        => 'date',
+            ]);
 
-        $userId = 2; //Id temporal, posteriormente se debe utilizar id de usuario en sesion
-        //Subir el archivo
-        if ($request->hasFile('_contT')) {
-            $file = $request->file('_contT');
+            //Subir el archivo imagen
+            if ($request->hasFile('_contT')) {
+                $file = $request->file('_contT');
+                
+                $filename = $usuarioId . '_2_' . $file->getClientOriginalName();
+
+                $filePath = $file->storeAs('uploads', $filename, 'public');
+            } else {
+                return redirect()->back()->with('error', 'Error al subir el archivo.');
+            }
             
-            $filename = $userId . '_2_' . $file->getClientOriginalName();
+            //Insert de publicación
+            $publicacion = new Publicaciones();
+            $publicacion->descripcion = $validator['_descT'];
+            $publicacion->nombre = $validator['_nt'];
+            $publicacion->contenido = $filePath;
+            $publicacion->fecha_publicacion = Carbon::now()->toDateString();
+            $publicacion->costo = $validator['_costoT'];
+            $publicacion->id_usuario = $usuarioId;
+            $publicacion->id_tipo = 2; //Tipo 2 porque son talleres
+            $publicacion->estatus = false;
+            $publicacion->save();
 
-            $filePath = $file->storeAs('uploads', $filename, 'public');
-        } else {
-            return redirect()->back()->with('error', 'Error al subir el archivo.');
+            return redirect()->back()->with('success', 'Publicación creada exitosamente.');
+        }
+        else {
+            return redirect('/');
         }
         
-        //Insert de publicación
-        $publicacion = new Publicaciones();
-        $publicacion->descripcion = $validator['_descT'];
-        $publicacion->nombre = $validator['_nt'];
-        $publicacion->contenido = $filePath;
-        $publicacion->fecha_publicacion = Carbon::now()->toDateString();
-        $publicacion->costo = $validator['_costoT'];
-        $publicacion->id_usuario = 2; //$validator['id_usuario']; //Temporal ya que aun no hay login
-        $publicacion->id_tipo = 2; 
-        $publicacion->estatus = true;//$validator['estatus']; //Temporal ya que aun no hay manejp de status
-        $publicacion->save();
-
-        return redirect()->back()->with('success', 'Publicación creada exitosamente.');
     }
 
     /**
@@ -101,46 +128,52 @@ class PublicacionesController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit($id)
+    public function edit()
     {
-        //
-        $publicacion = Publicaciones::findOrFail($id);
-        return view('editar_taller', compact('publicacion'));
+        $usuarioId = Auth::id();
+        $publicaciones = Publicaciones::where('id_usuario', $usuarioId)
+            ->with('usuario')
+            ->get();
+        return view('mis_talleres', compact('publicacion'));
     }
 
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, $id)
-    {
-        //
-        $publicacion = Publicaciones::findOrFail($id);
+    {        
+        if (Auth::check()){
+            //Busca todas los datos de dicha publicacion
+            $publicacion = Publicaciones::findOrFail($id);
 
-        // Validaciones
-        $validator = $request->validate([
-            '_descT' => 'required',
-            '_nt' => 'required|max:50',
-            '_contT' => 'nullable|file|max:2048', // Cambiado a nullable para permitir no cambiar el archivo
-            '_costoT' => 'numeric',
-        ]);
+            // Validaciones
+            $validator = $request->validate([
+                '_descT' => 'required',
+                '_nt' => 'required|max:50',
+                '_contT' => 'nullable|file|max:2048', // Cambiado a nullable para permitir no cambiar el archivo
+                '_costoT' => 'numeric',
+            ]);
 
-        // Actualización de campos
-        $publicacion->descripcion = $validator['_descT'];
-        $publicacion->nombre = $validator['_nt'];
-        $publicacion->costo = $validator['_costoT'];
+            // Actualización de campos
+            $publicacion->descripcion = $validator['_descT'];
+            $publicacion->nombre = $validator['_nt'];
+            $publicacion->costo = $validator['_costoT'];
 
-        // Subir el archivo si se proporciona uno nuevo
-        if ($request->hasFile('_contT')) {
-            $file = $request->file('_contT');
-            $filename = $publicacion->id_usuario . '_2_' . $file->getClientOriginalName();
-            $filePath = $file->storeAs('uploads', $filename, 'public');
-            $publicacion->contenido = $filePath;
+            // Subir el archivo si se proporciona uno nuevo
+            if ($request->hasFile('_contT')) {
+                $file = $request->file('_contT');
+                $filename = $publicacion->id_usuario . '_2_' . $file->getClientOriginalName();
+                $filePath = $file->storeAs('uploads', $filename, 'public');
+                $publicacion->contenido = $filePath;
+            }
+
+            // Guardar los cambios
+            $publicacion->save();
+            return redirect()->back();
         }
-
-        // Guardar los cambios
-        $publicacion->save();
-
-        return redirect()->back();
+        else{
+            return redirect('/');
+        }   
     }
 
     /**
@@ -148,16 +181,21 @@ class PublicacionesController extends Controller
      */
     public function physicalDestroy(string $id)
     {
-        //
-        $publicacion = Publicaciones::findOrFail($id);
-        if (Storage::exists('/public' . $publicacion->contenido)) {
-            Storage::delete('/public' . $publicacion->contenido);
+        if (Auth::check()){
+            //
+            $publicacion = Publicaciones::findOrFail($id);
+            if (Storage::exists('/public' . $publicacion->contenido)) {
+                Storage::delete('/public' . $publicacion->contenido);
+            }
+
+            //Eliminar publicación
+            $publicacion->delete();
+
+            return redirect()->back();
         }
+        else {
 
-        //Eliminar publicación
-        $publicacion->delete();
-
-        return redirect()->back();
+        }
     }
 
     // Definición de la relación con el modelo User
@@ -166,23 +204,28 @@ class PublicacionesController extends Controller
         return $this->belongsTo(Usuarios::class, 'id_usuario');
     }
 
-    //Borrad lógico de taller, desactiva la publicacion 
+    //Borrado lógico de taller, desactiva la publicacion 
     public function offStatus(Request $request, $id)
     {
-        $publicacion = Publicaciones::findOrFail($id);
-        $publicacion->estatus = false; // Cambia el estado a false
-        $publicacion->save();
-    
-        return redirect()->back();
+        if (Auth::check()){
+            $publicacion = Publicaciones::findOrFail($id);
+            $publicacion->estatus = false; // Cambia el estado a false
+            $publicacion->save();
+        
+            return redirect()->back();
+        }
     }
 
+    //Activacion de publicacion en caso de que este desactivada
     public function onStatus(Request $request, $id)
     {
-        $publicacion = Publicaciones::findOrFail($id);
-        $publicacion->estatus = true; // Cambia el estado a false
-        $publicacion->save();
-    
-        return redirect()->back();
+        if (Auth::check()){
+            $publicacion = Publicaciones::findOrFail($id);
+            $publicacion->estatus = true; // Cambia el estado a false
+            $publicacion->save();
+        
+            return redirect()->back();
+        }
     }
     
 }
